# fl_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/spiele": {
      "get": {
        "summary": "List Spiele",
        "description": "List Spiele matching the given filters.\n\nOmitting `saison_id` returns the CURRENT season. `saison_phase` accepts `playoffs` as an alias\nfor \"any phase except gruppenphase\", and `ausstehend` includes today.",
        "operationId": "get_spiele_api_v0_spiele_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "saison_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Id"
            }
          },
          {
            "name": "saison_phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "playoffs",
                  "type": "string"
                },
                {
                  "enum": [
                    "gruppenphase",
                    "achtelfinale",
                    "viertelfinale",
                    "halbfinale",
                    "finale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Phase"
            }
          },
          {
            "name": "spiel_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ausstehend",
                    "vergangen",
                    "heute",
                    "abgesagt",
                    "unbekannt"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Spiel Status"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1024,
              "minimum": 1,
              "default": 1024,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "datum",
                "uhrzeit",
                "spiel_nr",
                "saison_phase"
              ],
              "type": "string",
              "default": "datum",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpieleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/spiele/{spiel_id}": {
      "get": {
        "summary": "One Spiel",
        "description": "Return one match by its id.\n\nNo season is resolved and no status derived: the match carries its own `saison_id`, and\n`spiel_status` is a property of a query rather than of a match.",
        "operationId": "get_spiel_api_v0_spiele__spiel_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spiel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spiel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpieleSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a Spiel",
        "description": "Update one Spiel and resolve the season's bracket.\n\nThe payload is written wholesale: an omitted field is overwritten. A result can fill or empty the\nslots below it, each named in `advanced_to`. `dry_run=true` answers the same and writes nothing.",
        "operationId": "patch_spiel_data_api_v0_spiele__spiel_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spiel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spiel Id"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Report what this payload would move and destroy, and write nothing",
              "default": false,
              "title": "Dry Run"
            },
            "description": "Report what this payload would move and destroy, and write nothing"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPatchSpielDataPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPatchSpielDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v0/spieler": {
      "get": {
        "summary": "List Spieler",
        "description": "List players, normally for one team.\n\nUnlike the other resources, omitting `saison_id` does NOT resolve to the current season here --\ncallers narrow by `team_id` instead. Only `vorname` is guaranteed present on a player.",
        "operationId": "get_spieler_api_v0_spieler_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "name": "saison_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Id"
            }
          },
          {
            "name": "is_nachgetragen",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Nachgetragen"
            }
          },
          {
            "name": "stufe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "E1",
                    "E2",
                    "Q1",
                    "Q2",
                    "Q3",
                    "Q4"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stufe"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1024,
              "minimum": 1,
              "default": 1024,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "vorname",
                "nachname",
                "stufe",
                "nummer",
                "position"
              ],
              "type": "string",
              "default": "position",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Spieler",
        "description": "Create a player -- the person, and nothing else.\n\nThey belong to no team until they have a junction row, and no uniqueness rule applies to a name.",
        "operationId": "post_spieler_api_v0_spieler_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPostSpielerPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/spieler/{spieler_id}": {
      "get": {
        "summary": "One Spieler",
        "description": "Return one player -- the person only.\n\nNOT the flattened squad shape the list returns: those fields are season-scoped, and picking a\nseason here would make the answer depend on a default the caller never asked for.",
        "operationId": "get_spieler_by_id_api_v0_spieler__spieler_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spieler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spieler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a Spieler's name",
        "description": "Update a player's name. No fan-out: unlike a team or a venue, it is embedded in no other document.",
        "operationId": "patch_spieler_api_v0_spieler__spieler_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spieler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spieler Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPatchSpielerPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Retire a Spieler (soft delete)",
        "description": "Retire a player. SOFT: it stamps `inactive_since`, and their squad rows are LEFT ALONE.",
        "operationId": "delete_spieler_api_v0_spieler__spieler_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spieler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spieler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v0/spieler/memberships": {
      "get": {
        "summary": "Every Spieler with their squad rows",
        "description": "Every player, retired ones included, each with every squad row they hold.\n\n`GET /spieler` cannot answer it at any setting: with a `saison_id` the junction join is strict,\nwithout one a player with no row comes back missing `team_id`.",
        "operationId": "get_spieler_memberships_api_v0_spieler_memberships_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpielerMembershipsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FLPatchSpielDataPayload": {
        "properties": {
          "is_canceled": {
            "type": "boolean",
            "title": "Is Canceled"
          },
          "team1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamField"
              },
              {
                "type": "null"
              }
            ]
          },
          "team2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamField"
              },
              {
                "type": "null"
              }
            ]
          },
          "team1_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Quelle"
          },
          "team2_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Quelle"
          },
          "elfmeterschiessen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielElfmeterschiessen"
              },
              {
                "type": "null"
              }
            ]
          },
          "datum": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datum"
          },
          "uhrzeit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uhrzeit"
          },
          "ort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielOrtField"
              },
              {
                "type": "null"
              }
            ]
          },
          "schiedsrichter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielSchiedsrichterField"
              },
              {
                "type": "null"
              }
            ]
          },
          "notiz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notiz"
          }
        },
        "type": "object",
        "required": [
          "is_canceled",
          "team1",
          "team2",
          "team1_quelle",
          "team2_quelle",
          "elfmeterschiessen",
          "datum",
          "uhrzeit",
          "ort",
          "schiedsrichter",
          "notiz"
        ],
        "title": "FLPatchSpielDataPayload"
      },
...
      "FLSpieleListResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "spiele": {
            "items": {
              "$ref": "#/components/schemas/FLSpielJoined"
            },
            "type": "array",
            "title": "Spiele"
          }
        },
        "type": "object",
        "required": [
          "spiele"
        ],
        "title": "FLSpieleListResponse"
      },
      "FLSpieleSingleResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "spiel": {
            "$ref": "#/components/schemas/FLSpielJoined"
          }
        },
        "type": "object",
        "required": [
          "spiel"
        ],
        "title": "FLSpieleSingleResponse"
      },
...
      "FLSpielerListResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "spieler": {
            "items": {
              "$ref": "#/components/schemas/FLSpieler"
            },
            "type": "array",
            "title": "Spieler"
          }
        },
        "type": "object",
        "required": [
          "spieler"
        ],
        "title": "FLSpielerListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}